Put set_interval in a helper file
